pyproject.toml
[project]
name = "hypothesis-test"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.0",
    "wandb>=0.16.0",
    "openai>=1.0.0",
    "datasets>=2.16.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "aiohttp>=3.9.0",
]
